requirements.txt
pillow==11.3.0
rich==14.0.0
